NBAWebScrap.java
This file reads data from two directories given (latestRoster & editableRoster). This gets the latest rosters from the NBA and the player statistics and calculate a rating for every player. Users will be able to edit the editableRoster directory. The latestRoster directory will not be editable it will only stores the most current roster updates from the NBA.